queries/requirements.txt
mysql-connector-python
dotenv
django
pymysql